Offline contacts are now removed if garak-hide-offline-buddies is set.
This should have worked before but it turns out that contact-online-buddies
is not reliable: It continues to say 1 when it should be 0 - this is probably
a bug in the way secondary buddy node updates are handled by the elim daemon:
secondary updates (eg of the parent node) when the node is deleted are
currently suppressed, and it looks like the primary update of the parent node
occurs while the buddy is still flagged as online. Messy to fix, so we'll
work around it in garak for now.